refactor(profile): 个人中心与设置拆成两个独立页
首页改版第 1 阶段。原账户中心（AccountScreen）把身份、额度、设置分组
混在一页，且低频偏好还要再折叠一层「应用设置」子页；本次按职责切开：

- ProfileScreen（原 AccountScreen）：只留身份区、套餐用量、GitHub 入口，
  外加收藏 / 设置 / 关于我们三个入口与退出登录
- SettingsScreen（吸收原 AppSettingsScreen）：外观、界面语言、摘要语言、
  默认首页、邮件订阅、每日推送、外链打开方式、反馈，分三组平铺

设置不再与身份混排后，四个低频偏好可以直接展开，「应用设置」这层折叠随之
删除，路由 AppSettings 改名 Settings。「关于我们」与「设置」并列留在个人
中心——后续「我的」升为一级 tab 时，这两行一起挪进底栏扩展菜单。

埋点：设置页内各项事件名不变（语义未变）；个人中心的设置入口原为
settings_app_settings（指向应用设置子页），现改名 profile_open_settings。
